03-02-25/fix - adicionando dados via ssg, corrigindo arquitetura de pastas, nomeclatura de elementos e etc
