+++ mono-repo/projects/front-user/public/requisitos.md
@@ -39,64 +39,35 @@ OK - Desenvolver a seção de chamada!!!
 
 
 
-
-
-
-
-
-
       - Refatorar componente `PropertiesCarousel`
-        OK - Corrigir line-height dos textos, pois está cortando
-        OK - Corrigir versão mobile para que card seja responsivo ocupando toda a tela até `425px`
-        OK - Título e subtitulo deve adicionar `...` caso seja maior que o card `(Atualmente só o titulo tem)`
-        OK - Criar SSG com informações do card, como imagem, titulo e texto.
-        OK - Criar componente para o Card e aplicar o map dentro deles com os dados do SSG
-        OK - Utilizar estrutura html semântica no Card
-        OK - Corrigir piscada da imagem estourada ao recarregar a página
-        OK - Refatorar e reutilizar estilizações em comum
-
-        - AO CLICAR NO CARD, ELE LEVARA PARA UMA PÁGINA DE ACORDO COM O SLUG DO CARD GERANDO AS INFOS 
-        DO ANUNCIO EM DESTAQUE (FAZER FUTURAMENTE)
-        
-
-      * Refatorar StyledContainer
-        OK - Ele deve ser desestruturado em vez de ser feito `styled(StyledContainer)`
-        OK - Deve ter outra nomeclatura como `StyledGlobalContainer` ou algo do tipo
-        OK - Os paddings gerais de bordas devem ser desestruturados para simplificar
-
-
+         - AO CLICAR NO CARD, ELE LEVARA PARA UMA PÁGINA DE ACORDO COM O SLUG DO CARD GERANDO AS INFOS DO ANUNCIO EM DESTAQUE (FAZER FUTURAMENTE)
+      
 
         
       * Criar seção `ElegantShowcase`
         OK - Criar conteúdo textual e adicionar estilização
+        OK - Implementar SSG
         - Criar player e adicionar video que está em /public/video
-        - Implementar SSG
 
 
       * Criar seção `TestimonialCarousel`
+        `AO CLICAR NO CARD, ELE LEVARA PARA A PÁGINA DO USUÁRIO QUE FEZ O COMENTÁRIO (FAZER FUTURAMENTE)`
         OK - Reutilizar Carousel do `PropertiesCarousel`
         OK - Criar conteúdo textual e adicionar estilização
         OK - Implementar SSG
-        - Reutilizar estilos do `PropertiesCard`
-        - AO CLICAR NO CARD, ELE LEVARA PARA A PÁGINA DO USUÁRIO QUE FEZ O COMENTÁRIO (FAZER FUTURAMENTE)
+        OK - Reutilizar estilos do `PropertiesCard`
 
       * Criar `Footer`
+        `O email cadastrado no final da página, será enviado para uma plataforma de marketing de email que ficara responsável por automatizar o envio de artigos criados pelo dono do site.`
         OK - Adicionar elementos do footer
         OK - adicionar input e social midia
         OK - adicionar mensagem de direitos
-        - disponibilizar dados por ssg se necessário
-        - O email cadastrado no final da página, será enviado para uma plataforma de marketing de email
-      que ficara responsável por automatizar o envio de artigos criados pelo dono do site.
+        OK - disponibilizar dados por ssg se necessário
 
-
-      * Conferir se devo utilizar SSG em cada componente ou Centralizado na index.tsx
-
-      * Integrar API de CADASTRO E LOGIN
-
-      * Refatorar arquitetura das INTERFACES de Tipagem
-      * Refatorar arquitetura das `Data` do `HomePage` (Não há necessidade de uma pasta pra cada data)
-      * Refatorar arquitetura da `index.tsx` criar uma `screen` da `HomePage` e importar na `index.tsx`
-      * Criar possível forma de reutilização do estilo da `StyledMainContainer`.
+        OK - Refatorar arquitetura das INTERFACES de Tipagem
+        OK - Refatorar arquitetura das `Data` do `HomePage` (Não há necessidade de uma pasta pra cada data)
+        OK - Refatorar arquitetura da `index.tsx` criar uma `screen` da `HomePage` e importar na `index.tsx`
+        OK - Criar possível forma de reutilização do estilo da `StyledMainContainer`.
 
 
       * Adicionar verificação de identação e etc...
