+++ mono-repo/projects/front-user/src/components/Header/AuthForm/RegistrationForm/RegistrationInputs/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import InputBox from 'src/components/InputBox';
-import { StyledInput } from 'src/styles/styledComponents/StyledForm';
+import { StyledInput } from 'src/styles/components/StyledForm';
 import UserIcon from 'public/icons/user.svg';
 import EmailIcon from 'public/icons/email.svg';
 import SecurityIcon from 'public/icons/security.svg';
