+++ mono-repo/projects/front-user/src/data/homePage/homePageData.ts
@@ -1,0 +1,17 @@
+import menuLinks from './json/menuLinks.json';
+import dropdowns from './json/dropdowns.json';
+import heroSection from './json/heroSection.json';
+import properties from './json/properties.json';
+import testimonials from './json/testimonials.json';
+import elegantShowcase from './json/elegantShowcase.json';
+import footer from './json/footer.json';
+
+export const homePageData = {
+  menuLinks,
+  dropdowns,
+  heroSection,
+  properties,
+  testimonials,
+  elegantShowcase,
+  footer,
+};
