+++ mono-repo/projects/front-user/src/data/homePage/json/elegantShowcase.json
@@ -1,0 +1,4 @@
+{
+  "title": "Viva a Elegância em \n Cada Detalhe",
+  "description": "Apresentamos uma coleção única de imóveis com design inteligente, acabamentos de alta qualidade e situados nas melhores áreas. Eleve seu estilo de vida com o que há de mais sofisticado no mercado imobiliário."
+}
