+++ mono-repo/projects/front-user/src/data/homePage/json/footer.json
@@ -1,0 +1,4 @@
+{
+  "title": "Inscreva-se para receber as melhores oportunidades e tendências no setor imobiliário direto no seu e-mail!",
+  "description": "© Todos os direitos reservados."
+}
