+++ mono-repo/projects/front-user/src/data/homePage/json/heroSection.json
@@ -1,4 +1,4 @@
 {
-  "titulo": "KondoRu",
+  "title": "KondoRu",
   "slogan": "Viva o melhor da vida\nno lugar certo"
 }
