+++ mono-repo/projects/front-user/src/data/homePage/json/menuLinks.json
