+++ mono-repo/projects/front-user/src/interfaces/homePage/data/IAuthForm.ts
