+++ mono-repo/projects/front-user/src/interfaces/homePage/data/IElegantShowcase.ts
@@ -1,0 +1,4 @@
+export interface IElegantShowCase {
+  title: string,
+  description: string
+}
