+++ mono-repo/projects/front-user/src/interfaces/homePage/data/IFooter.ts
@@ -1,0 +1,5 @@
+export interface IFooter {
+  title: string,
+  description: string,
+  year: number
+}
