+++ mono-repo/projects/front-user/src/interfaces/homePage/data/IHeroSection.ts
@@ -1,0 +1,4 @@
+export interface IHeroSection {
+  title: string;
+  slogan: string;
+}
