+++ mono-repo/projects/front-user/src/interfaces/homePage/data/IPropertyCard.ts
