+++ mono-repo/projects/front-user/src/interfaces/homePage/data/ITestimonialCard.ts
