+++ mono-repo/projects/front-user/src/screens/HomeScreen/ElegantShowcase/ElegantShowCasePlayer/index.tsx
