+++ mono-repo/projects/front-user/src/screens/HomeScreen/PropertyCarousel/index.tsx
@@ -1,7 +1,7 @@
-import { SwiperSlide } from 'swiper/react';
 import { useHomePageData } from 'src/contexts/HomePageContext';
-import Carousel from '../Carousel';
+import { SwiperSlide } from 'swiper/react';
 import PropertyCard from './PropertyCard';
+import Carousel from 'src/components/Carousel';
 
 export default function PropertyCarousel() {
   const { properties } = useHomePageData();
