+++ mono-repo/projects/front-user/src/styles/components/StyledForm.ts
