+++ mono-repo/projects/front-user/src/styles/components/StyledLayoutWrapper.ts
