+++ mono-repo/projects/front-user/src/styles/components/StyledMain.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const StyledMainContainer = styled.main`
+  display: flex;
+  flex-direction: column;
+  gap: 6.25rem;
+  padding-bottom: 6.25rem;
+
+  @media (min-width: 768px) {
+    gap: 8rem;
+    padding-bottom: 8rem;
+  }
+`;
