+++ mono-repo/projects/front-user/src/styles/components/StyledText.ts
