+++ mono-repo/projects/front-user/src/styles/font.ts
@@ -2,7 +2,7 @@ import { Mulish } from 'next/font/google';
 
 const mulish = Mulish({
   subsets: ['latin'],
-  weight: ['500', '600', '700', '800', '900'],
+  weight: ['300', '400', '500', '600', '700', '800', '900'],
   display: 'swap',
 });
